refactor: SessionType 작업
